✨ feat: add files and centralize route config
Add robots.txt and sitemap metadata routes and extract route
configuration into lib/constants/routes.ts to centralize public,
public-only, and protected route logic. Move route-checking helpers
(isPublicRoute, isProtectedRoute) out of Supabase middleware and
import the shared functions there to avoid duplication and ensure a
single source of truth for auth-related route rules.

Update middleware patterns to allow sitemap.xml and robots.txt to be
served directly (exclude them from auth/middleware handling) and
document excluded static paths. These changes improve SEO support
(sitemap/robots) and simplify maintenance of route access rules for
authentication and sitemap generation.
