+++ app/robots.ts
@@ -1,0 +1,18 @@
+import { MetadataRoute } from "next";
+
+/**
+ * Generate robots.txt for SEO
+ * Allows all pages for crawling - actual access control handled by middleware
+ * @see https://nextjs.org/docs/app/api-reference/file-conventions/metadata/robots
+ */
+export default function robots(): MetadataRoute.Robots {
+  const baseUrl = process.env.NEXT_PUBLIC_SITE_URL!;
+
+  return {
+    rules: {
+      userAgent: "*",
+      allow: "/",
+    },
+    sitemap: `${baseUrl}/sitemap.xml`,
+  };
+}
