+++ lib/supabase/middleware.ts
@@ -1,61 +1,6 @@
 import { createServerClient } from "@supabase/ssr";
 import { NextResponse, type NextRequest } from "next/server";
-
-interface Routes {
-  [key: string]: boolean;
-}
-const publicRoutes: Routes = {
-  "/login": true,
-  "/register": true,
-  "/register/success": true,
-  "/auth": true,
-  "/error": true,
-  "/": true,
-  "/forgot-password": true,
-  "/reset-password": true,
-  "/privacy": true,
-  "/term": true,
-  "/contact": true,
-};
-
-// Routes that should redirect to dashboard if user is authenticated
-const publicOnlyRoutes: Routes = {
-  "/login": true,
-  "/register": true,
-  "/register/success": true,
-  "/": true,
-};
-
-// Function to check if a path is public
-function isPublicRoute(pathname: string): boolean {
-  // Check explicit public routes
-  if (publicRoutes[pathname]) return true;
-
-  // Check if it's a sign route (public for external users)
-  if (pathname.startsWith("/sign/")) return true;
-
-  // Check if it's an auth route (including auth/confirm, auth/callback, etc.)
-  if (pathname.startsWith("/auth/")) return true;
-
-  return false;
-}
-
-// Protected routes that require authentication
-const protectedRoutes: Routes = {
-  "/upload": true,
-  "/document": true,
-  "/pricing": true,
-};
-
-// Function to check if a path is protected
-function isProtectedRoute(pathname: string): boolean {
-  // Check if the path starts with any protected route pattern
-  return (
-    Object.keys(protectedRoutes).some(
-      (route) => pathname.startsWith(route) || pathname === route
-    ) || pathname.startsWith("/document/")
-  );
-}
+import { publicOnlyRoutes, isPublicRoute } from "@/lib/constants/routes";
 
 export async function updateSession(request: NextRequest) {
   let supabaseResponse = NextResponse.next({
