+++ middleware.ts
@@ -12,9 +12,11 @@ export const config = {
      * - _next/static (static files)
      * - _next/image (image optimization files)
      * - favicon.ico (favicon file)
+     * - sitemap.xml (SEO sitemap)
+     * - robots.txt (SEO robots)
      * - .*\\.(?:ico|png|svg|jpg|jpeg|gif|webp)$ (image files)
      * - api/webhook (Paddle webhook endpoint)
      */
-    "/((?!_next/static|_next/image|favicon.ico|api/webhook|.*\\.(?:ico|png|svg|jpg|jpeg|gif|webp)$).*)",
+    "/((?!_next/static|_next/image|favicon.ico|sitemap.xml|robots.txt|api/webhook|.*\\.(?:ico|png|svg|jpg|jpeg|gif|webp)$).*)",
   ],
 };
